extract current cpu chart into separate component
