+++ frontend/src/pages/instance/routes.js
@@ -8,7 +8,7 @@ import InstanceRemoteInstances from './Devices.vue'
 import DuplicateInstance from './DuplicateInstance.vue'
 import InstanceLogs from './Logs.vue'
 import InstanceOverview from './Overview.vue'
-import InstancePerformance from './Performance.vue'
+import InstancePerformance from './Performance/index.vue'
 import InstanceSettings from './Settings/index.vue'
 import InstanceSettingsRoutes from './Settings/routes.js'
 import VersionHistory from './VersionHistory/index.vue'
